basic porting support for qmk

+++ keyboards/onekey/onekey_keymap.c
@@ -12,6 +12,6 @@
 const action_t fn_actions[] = {
 };
 
-const uint8_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
+const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
     [0] = LAYOUT_ext(KC_ESC, KC_1, KC_2, KC_3, KC_4, KC_5)
 };
